@@ -28,12 +28,12 @@ public class AssortmentService {
         if (assortments != null && !assortments.isEmpty()) {
             return assortments;
         } else {
-            LOG.debug("Value for shop=" + shop.getTitle() + " not found in cache. He will be updated.");
+            LOG.info("Value for shop=" + shop.getTitle() + " not found in cache. He will be updated.");
 
             assortments = db.getAssortments(shop);
             cacheAssortment.computeIfAbsent(shop, k -> new ArrayList<>())
                     .addAll(assortments);
-            return getAssortments(shop);
+            return assortments;
         }
     }
 
